// Frontend/src/app/api-client/model/createTaskRequest.ts
/**
 * Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateTaskRequest { 
    title?: string | null;
    description?: string | null;
    dueDate?: string;
    isCompleted?: boolean;
}